+++ packages/cheap-di-ts-transform/src/generation/callFunction.ts
@@ -6,7 +6,16 @@ import { makeIdentifier } from './makeIdentifier.js';
  * @example
  * yourFunctionName(parameter1, parameter2);
  * */
-export function callFunction(
+export function callFunction(...functionParameters: Parameters<typeof callFunctionExpression>) {
+  return ts.factory.createExpressionStatement(callFunctionExpression(...functionParameters));
+}
+
+/**
+ * generates function expression to using to using in another statements
+ * @example
+ * yourFunctionName(parameter1, parameter2)
+ * */
+export function callFunctionExpression(
   funcOrNameOrIdentifier: string | ts.Identifier | ts.PropertyAccessExpression,
   ...functionParameters: ts.Expression[]
 ) {