+++ packages/cheap-di-ts-transform/src/generation/immediateFunction.ts
@@ -1,0 +1,11 @@
+import ts from 'typescript';
+
+/**
+ * @example
+ * (<your function>)();
+ * */
+export function immediateFunction(functionExpression: ts.Expression) {
+  return ts.factory.createExpressionStatement(
+    ts.factory.createCallExpression(ts.factory.createParenthesizedExpression(functionExpression), undefined, undefined)
+  );
+}