+++ packages/cheap-di-ts-transform/src/generation/requireImport.ts
@@ -1,0 +1,10 @@
+import { callFunctionExpression } from './callFunction.js';
+import ts from 'typescript';
+
+/**
+ * @example
+ * require('<imported from>')
+ * */
+export function requireImport(importedFrom: string) {
+  return callFunctionExpression('require', ts.factory.createStringLiteral(importedFrom));
+}