+++ tests/package1/src/Service1.ts
@@ -1,5 +1,0 @@
-export class Service1 {
-  log(message: string) {
-    console.log(message);
-  }
-}