+++ tests/vite-project/vite.config.ts
@@ -1,0 +1,30 @@
+import { defineConfig } from 'vite';
+import react from '@vitejs/plugin-react';
+import typescript from '@rollup/plugin-typescript';
+import { transformer } from 'cheap-di-ts-transform';
+
+export default defineConfig({
+  plugins: [
+    react(),
+    typescript({
+      transformers: {
+        before: [
+          {
+            type: 'program',
+            factory: (program) =>
+              transformer(
+                { program },
+                {
+                  debug: true,
+                  addDetailsToUnknownParameters: true,
+                  logRegisteredMetadata: true,
+                  errorsLogLevel: 'debug',
+                  esmImports: true,
+                }
+              ),
+          },
+        ],
+      },
+    }),
+  ],
+});