+++ packages/cheap-di/CHANGELOG.md
@@ -1,6 +1,118 @@
 # Changelog
 
-### 3.5.0
+## 4.0.0
+
+The release main goal is stop using reflect-metadata as obsoleted (and unstable) approach and become compatible with TypeScript 5.2.0 decorators (stage 3 proposal).
+
+As a bonus we introduce new approach to dependency registration to achieve True dependency injection - <a href="https://github.com/tomas-light/cheap-di/blob/master/packages/cheap-di-ts-transform/README.md">cheap-di-ts-transform</a> package
+
+### 🚧 Breaking changes
+
+* `registerType` method was renamed to `registerImplementation`;
+
+```ts
+import { container } from 'cheap-di';
+
+class Foo {}
+
+// before
+container.registerType(Foo);
+
+// after
+container.registerImplementation(Foo);
+```
+
+* parameters injection method was renamed;
+
+```ts
+import { container } from 'cheap-di';
+
+class Foo {
+  constructor(bar: string) {}
+}
+
+// before
+container
+  .registerType(Foo)
+  .with('some message');
+
+// after
+container
+  .registerImplementation(Foo)
+  .inject('some message');
+```
+
+* `@inject` decorator was rewritten and its target changed from constructor parameter to constructor itself:
+
+```ts
+import { inject } from 'cheap-di';
+
+class Bar {}
+
+// before
+class Foo {
+  constructor(@inject bar: Bar) {}
+}
+
+// after
+@inject(Bar)
+class Foo {
+  constructor(bar: Bar) {}
+}
+```
+
+* removed `@dependencies` decorator. You may use `@inject` decorator instead of;
+* removed `@di` decorator;
+* removed `@singleton` decorator. Instead of it you may use:
+
+```ts
+import { container } from 'cheap-di';
+
+class Foo {}
+
+// here you have to provide Foo's dependencies by yourself
+container.registerInstance(new Foo());
+
+// or
+
+// Foo will be instantiated only once, all consumers will get the same instance
+container
+  .registerImplementation(Foo)
+  .asSingleton(); 
+```
+
+* All symbols `dependenciesSymbol`, `singletonSymbol`, `injectionSymbol`, `inheritancePreserveSymbol` were removed, all information now are stored in single `cheapDiSymbol` symbol:
+
+```ts
+type ImplementationType<TClass> = Dependency<TClass> & {
+  [cheapDiSymbol]: DiMetadata;
+};
+
+interface DiMetadata {
+  /** if class was registered as singleton with container.registerImplementation(...).asSingleton)*/
+  singleton?: boolean;
+
+  /** constructor dependencies */
+  dependencies?: SomeDependency[];
+
+  /** reference on a constructor that was patched (helps in inheritance cases) */
+  modifiedClass?: unknown;
+
+  /** parameters are specified with @inject decorator explicitly */
+  injectDependencies?: SomeDependency[];
+
+  /** parameters are passed to container.registerImplementation(...).inject(parameter1, parameter2) */
+  injected?: unknown[];
+} 
+```
+
+### 🚀 Features:
+
+* `@inject` decorator supports stage 2 and stage 3TypeScript decorator syntax;
+* get rid of reflect-metadata issues, when your class loses its dependencies in runtime in unpredictable cases;
+* provided separated bundles for CommonJs and EcmaScript modules;
+
+## 3.5.0
 
 🐛 fix tslib missing dependency for PNP
 
@@ -9,23 +121,23 @@
 🔨 bump dependencies;
 
 
-### 3.4.3
+## 3.4.3
 
 🔨 Bump dependencies versions (and typescript from v4.7.3 to v4.9.4)
 
-### 3.4.2
+## 3.4.2
 
 Types: 
 
 🔨 changed access level of `getInstance`, `getImplementation`, `getSingletons` methods in `ContainerImpl` to public;
 
-### 3.4.1
+## 3.4.1
 
 Types:
 
 🔨 `container` now is typed as `ContainerImpl`;
 
-### 3.4.0
+## 3.4.0
 
 Types:
 
@@ -33,33 +145,33 @@ Types:
 
 🔨 `ContainerImpl` implements new interfaces to be able to implement them in your container;
 
-### 3.3.2
+## 3.3.2
 
 Bugfixes:
 
 🐛 try to get injected params in prototype chain if not found anything in current constructor during resolve;
 
-### 3.3.1
+## 3.3.1
 
 Bugfixes:
 
 🐛 bugfix of dependency auto resolving;
 
-### 3.2.5
+## 3.2.5
 
 Bugfixes:
 
 🐛 revert building for CommonJS;
 
-### 3.2.4
+## 3.2.4
 
 🔨 update dependencies;
 
-### 3.2.3
+## 3.2.3
 
 🔨 fix sources map
 
-### 3.2.1
+## 3.2.1
 
 Bugfixes:
 
@@ -69,11 +181,11 @@ Bugfixes:
 
 🐛 fix auto resolving for inherited classes;
 
-### 3.2.0
+## 3.2.0
 
 🚀 add `asSingleton` registration method;
 
-### 3.1.0
+## 3.1.0
 Features:
 
 🚀 add `@di` class-decorator;
@@ -82,7 +194,7 @@ Features:
 
 🚀 add `clear` method to `DependencyResolver`;
 
-### 3.0.0
+## 3.0.0
 
 * move `parentContainer` field to `cheap-di-react`;
 * move `sameParent` method to `cheap-di-react`;
@@ -93,7 +205,7 @@ Features:
 * change access to `getImplementation` method by `protected`;
 * change access to `getSingletons` method by `protected`;
 
-### 2.3.0
+## 2.3.0
 
 * prevent decorated setting inheritance (because it's a static field)
 * add `getDependencies` method to get settled types
@@ -106,11 +218,11 @@ Features:
 * add `setInjectedParams` method to set params (it's not `@inject` analog), that will be passed after all dependencies in your class
 * add `getInjectedParams` method to get injected params
 
-### 2.2.3
+## 2.2.3
 
 * bugfix using of `@inject` in inherited classes
 
-### 2.2.2
+## 2.2.2
 
 * bugfix singletons -> move all singleton registration to root container
 * improve instance resolving
@@ -129,11 +241,11 @@ class New {
 ```
 * add error for circular dependencies with dependency trace
 
-### 2.2.1
+## 2.2.1
 
 * add `@inject` to index import
 
-### 2.2.0
+## 2.2.0
 
 * replace static property names by Symbols:
 ```ts
@@ -177,29 +289,29 @@ class MyClass {
 }
 ```
 
-### 2.1.0
+## 2.1.0
 
 * add `@singleton` class-decorator
 
-### 2.0.5
+## 2.0.5
 
 * fix type error: for class that contains static members
 * fix type error: for type registration with injection params
 * fix type error: add abstract constructor for instance registration
 * add className property for decorated classes
 
-### 2.0.1
+## 2.0.1
 
 * add Container instantiation (with container nesting)
 * add 'dependencies' decorator for TypeScript classes
 * bugfixes
 
-### 2.0.0
+## 2.0.0
 
 * rename RegisteredType to RegistrationType
 * rename __constructorParams to __dependencies
 
-### 1.1.0
+## 1.1.0
 
 * rename IContainer to Container
 * rename IDependencyRegistrator to DependencyRegistrator